requirements.txt
aiohttp
aiosignal
amqp
asgiref
async-timeout
attrs
azure-core
azure-storage-blob
billiard
boto3
botocore
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
crontab
cryptography
Django
django-cors-headers
django-rq
django-storages
djangorestframework
drf-spectacular
freezegun
frozenlist
gunicorn
idna
inflection
isodate
jmespath
jsonschema
jsonschema-specifications
kombu
multidict
numpy
openai
opencv-python-headless
packaging
prompt-toolkit
psycopg2-binary
pycparser
python-dateutil
python-dotenv
pytz
PyYAML
redis
referencing
requests
rq
rq-scheduler
s3transfer
six
sqlparse
tqdm
typing_extensions
tzdata
uritemplate
urllib3
vine
wcwidth
yarl
